@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as eslint from 'eslint';
+import type * as ESTree from 'estree';
 import { TSESTree } from '@typescript-eslint/utils';
 
 const VALID_USES = new Set<TSESTree.AST_NODE_TYPES | undefined>([
@@ -11,22 +12,22 @@ const VALID_USES = new Set<TSESTree.AST_NODE_TYPES | undefined>([
 	TSESTree.AST_NODE_TYPES.VariableDeclarator,
 ]);
 
-export = new class MustUseResults implements eslint.Rule.RuleModule {
+export default new class MustUseResults implements eslint.Rule.RuleModule {
 	readonly meta: eslint.Rule.RuleMetaData = {
 		schema: false
 	};
 
 	create(context: eslint.Rule.RuleContext): eslint.Rule.RuleListener {
 
-		const config = <{ message: string; functions: string[] }[]>context.options[0];
+		const config = context.options[0] as { message: string; functions: string[] }[];
 		const listener: eslint.Rule.RuleListener = {};
 
 		for (const { message, functions } of config) {
 			for (const fn of functions) {
 				const query = `CallExpression[callee.property.name='${fn}'], CallExpression[callee.name='${fn}']`;
-				listener[query] = (node: any) => {
-					const cast: TSESTree.CallExpression = node;
-					if (!VALID_USES.has(cast.parent?.type)) {
+				listener[query] = (node: ESTree.Node) => {
+					const callExpression = node as TSESTree.CallExpression;
+					if (!VALID_USES.has(callExpression.parent?.type)) {
 						context.report({ node, message });
 					}
 				};
