@@ -4,6 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as eslint from 'eslint';
+import type * as ESTree from 'estree';
+import { TSESTree } from '@typescript-eslint/utils';
 
 /**
  * Disallows the use of the `in` operator in TypeScript code, except within
@@ -15,7 +17,7 @@ import * as eslint from 'eslint';
  * Exception: Type predicate functions are allowed to use the `in` operator
  * since they are the standard way to perform runtime type checking.
  */
-export = new class NoInOperator implements eslint.Rule.RuleModule {
+export default new class NoInOperator implements eslint.Rule.RuleModule {
 
 	readonly meta: eslint.Rule.RuleMetaData = {
 		messages: {
@@ -26,9 +28,10 @@ export = new class NoInOperator implements eslint.Rule.RuleModule {
 
 	create(context: eslint.Rule.RuleContext): eslint.Rule.RuleListener {
 
-		function checkInOperator(inNode: any) {
+		function checkInOperator(inNode: ESTree.BinaryExpression) {
+			const node = inNode as TSESTree.BinaryExpression;
 			// Check if we're inside a type predicate function
-			const ancestors = context.sourceCode.getAncestors(inNode);
+			const ancestors = context.sourceCode.getAncestors(node as ESTree.Node);
 
 			for (const ancestor of ancestors) {
 				if (ancestor.type === 'FunctionDeclaration' ||
@@ -45,7 +48,7 @@ export = new class NoInOperator implements eslint.Rule.RuleModule {
 			}
 
 			context.report({
-				node: inNode,
+				node,
 				messageId: 'noInOperator'
 			});
 		}
