@@ -4,8 +4,9 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as eslint from 'eslint';
+import type * as ESTree from 'estree';
 
-export = new class ApiProviderNaming implements eslint.Rule.RuleModule {
+export default new class ApiProviderNaming implements eslint.Rule.RuleModule {
 
 	readonly meta: eslint.Rule.RuleMetaData = {
 		messages: {
@@ -17,13 +18,13 @@ export = new class ApiProviderNaming implements eslint.Rule.RuleModule {
 	create(context: eslint.Rule.RuleContext): eslint.Rule.RuleListener {
 
 		return {
-			['PropertyDefinition PrivateIdentifier']: (node: any) => {
+			['PropertyDefinition PrivateIdentifier']: (node: ESTree.Node) => {
 				context.report({
 					node,
 					messageId: 'slow'
 				});
 			},
-			['MethodDefinition PrivateIdentifier']: (node: any) => {
+			['MethodDefinition PrivateIdentifier']: (node: ESTree.Node) => {
 				context.report({
 					node,
 					messageId: 'slow'
