@@ -4,23 +4,24 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as eslint from 'eslint';
+import type * as ESTree from 'estree';
 
 /**
  * Checks for potentially unsafe usage of `DisposableStore` / `MutableDisposable`.
  *
  * These have been the source of leaks in the past.
  */
-export = new class implements eslint.Rule.RuleModule {
+export default new class implements eslint.Rule.RuleModule {
 
 	create(context: eslint.Rule.RuleContext): eslint.Rule.RuleListener {
-		function checkVariableDeclaration(inNode: any) {
+		function checkVariableDeclaration(inNode: ESTree.Node) {
 			context.report({
 				node: inNode,
 				message: `Use const for 'DisposableStore' to avoid leaks by accidental reassignment.`
 			});
 		}
 
-		function checkProperty(inNode: any) {
+		function checkProperty(inNode: ESTree.Node) {
 			context.report({
 				node: inNode,
 				message: `Use readonly for DisposableStore/MutableDisposable to avoid leaks through accidental reassignment.`
