@@ -5,11 +5,12 @@
 
 import { TSESTree } from '@typescript-eslint/utils';
 import * as eslint from 'eslint';
+import type * as ESTree from 'estree';
 
-export = new class NoReaderAfterAwait implements eslint.Rule.RuleModule {
+export default new class NoReaderAfterAwait implements eslint.Rule.RuleModule {
 	create(context: eslint.Rule.RuleContext): eslint.Rule.RuleListener {
 		return {
-			'CallExpression': (node: any) => {
+			'CallExpression': (node: ESTree.CallExpression) => {
 				const callExpression = node as TSESTree.CallExpression;
 
 				if (!isFunctionWithReader(callExpression.callee)) {
