@@ -3,8 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { TSESTree } from '@typescript-eslint/utils';
 import * as eslint from 'eslint';
+import type * as ESTree from 'estree';
+import { TSESTree } from '@typescript-eslint/utils';
 
 function isCallExpression(node: TSESTree.Node): node is TSESTree.CallExpression {
 	return node.type === 'CallExpression';
@@ -14,13 +15,14 @@ function isFunctionExpression(node: TSESTree.Node): node is TSESTree.FunctionExp
 	return node.type.includes('FunctionExpression');
 }
 
-export = new class NoAsyncSuite implements eslint.Rule.RuleModule {
+export default new class NoAsyncSuite implements eslint.Rule.RuleModule {
 
 	create(context: eslint.Rule.RuleContext): eslint.Rule.RuleListener {
-		function hasAsyncSuite(node: any) {
-			if (isCallExpression(node) && node.arguments.length >= 2 && isFunctionExpression(node.arguments[1]) && node.arguments[1].async) {
+		function hasAsyncSuite(node: ESTree.Node) {
+			const tsNode = node as TSESTree.Node;
+			if (isCallExpression(tsNode) && tsNode.arguments.length >= 2 && isFunctionExpression(tsNode.arguments[1]) && tsNode.arguments[1].async) {
 				return context.report({
-					node: node,
+					node: tsNode,
 					message: 'suite factory function should never be async'
 				});
 			}
