@@ -3,10 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as eslint from 'eslint';
 import { AST_NODE_TYPES, TSESTree } from '@typescript-eslint/utils';
+import * as eslint from 'eslint';
 
-export = new class ApiProviderNaming implements eslint.Rule.RuleModule {
+export default new class ApiProviderNaming implements eslint.Rule.RuleModule {
 
 	readonly meta: eslint.Rule.RuleMetaData = {
 		messages: {
@@ -18,10 +18,10 @@ export = new class ApiProviderNaming implements eslint.Rule.RuleModule {
 	create(context: eslint.Rule.RuleContext): eslint.Rule.RuleListener {
 
 		return {
-			['TSInterfaceDeclaration[id.name=/.+Provider/] TSMethodSignature[key.name=/^(provide|resolve).+/]']: (node: any) => {
+			['TSInterfaceDeclaration[id.name=/.+Provider/] TSMethodSignature[key.name=/^(provide|resolve).+/]']: (node: TSESTree.Node) => {
 
 				let found = false;
-				for (const param of (<TSESTree.TSMethodSignature>node).params) {
+				for (const param of (node as TSESTree.TSMethodSignature).params) {
 					if (param.type === AST_NODE_TYPES.Identifier) {
 						found = found || param.name === 'token';
 					}
