@@ -4,9 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as eslint from 'eslint';
+import type * as ESTree from 'estree';
 import { TSESTree } from '@typescript-eslint/utils';
 
-export = new class ApiInterfaceNaming implements eslint.Rule.RuleModule {
+export default new class ApiInterfaceNaming implements eslint.Rule.RuleModule {
 
 	private static _nameRegExp = /^I[A-Z]/;
 
@@ -20,9 +21,9 @@ export = new class ApiInterfaceNaming implements eslint.Rule.RuleModule {
 	create(context: eslint.Rule.RuleContext): eslint.Rule.RuleListener {
 
 		return {
-			['TSInterfaceDeclaration Identifier']: (node: any) => {
+			['TSInterfaceDeclaration Identifier']: (node: ESTree.Identifier) => {
 
-				const name = (<TSESTree.Identifier>node).name;
+				const name = (node as TSESTree.Identifier).name;
 				if (ApiInterfaceNaming._nameRegExp.test(name)) {
 					context.report({
 						node,
