@@ -3,10 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as eslint from 'eslint';
 import { TSESTree } from '@typescript-eslint/utils';
+import * as eslint from 'eslint';
 
-export = new class ApiLiteralOrTypes implements eslint.Rule.RuleModule {
+export default new class ApiLiteralOrTypes implements eslint.Rule.RuleModule {
 
 	readonly meta: eslint.Rule.RuleMetaData = {
 		docs: { url: 'https://github.com/microsoft/vscode/wiki/Extension-API-guidelines#enums' },
@@ -16,8 +16,8 @@ export = new class ApiLiteralOrTypes implements eslint.Rule.RuleModule {
 
 	create(context: eslint.Rule.RuleContext): eslint.Rule.RuleListener {
 		return {
-			['TSTypeAnnotation TSUnionType']: (node: any) => {
-				if ((<TSESTree.TSUnionType>node).types.every(value => value.type === 'TSLiteralType')) {
+			['TSTypeAnnotation TSUnionType']: (node: TSESTree.TSUnionType) => {
+				if (node.types.every(value => value.type === 'TSLiteralType')) {
 					context.report({
 						node: node,
 						messageId: 'useEnum'
