@@ -82,7 +82,7 @@ struct HandlerContext {
 	socket_tx: mpsc::Sender<SocketSignal>,
 	/// Configured launcher paths.
 	launcher_paths: LauncherPaths,
-	/// Connected Notepad# Server
+	/// Connected VS Code Server
 	code_server: CodeServerCell,
 	/// Potentially many "websocket" connections to client
 	server_bridges: ServerMultiplexer,
@@ -90,7 +90,7 @@ struct HandlerContext {
 	code_server_args: CodeServerArgs,
 	/// port forwarding functionality
 	port_forwarding: Option<PortForwarding>,
-	/// install platform for the Notepad# server
+	/// install platform for the VS Code server
 	platform: Platform,
 	/// http client to make download/update requests
 	http: Arc<FallbackSimpleHttp>,
