@@ -2,13 +2,13 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-// @ts-nocheck
+// @ts-check
 import fs from 'fs';
 import path from 'path';
 import tseslint from 'typescript-eslint';
 
 import stylisticTs from '@stylistic/eslint-plugin-ts';
-import * as pluginLocal from './.eslint-plugin-local/index.js';
+import * as pluginLocal from './.eslint-plugin-local/index.ts';
 import pluginJsdoc from 'eslint-plugin-jsdoc';
 
 import pluginHeader from 'eslint-plugin-header';
@@ -75,7 +75,7 @@ export default tseslint.config(
 				'context'
 			], // non-complete list of globals that are easy to access unintentionally
 			'no-var': 'warn',
-			'semi': 'off',
+			'semi': 'warn',
 			'local/code-translation-remind': 'warn',
 			'local/code-no-native-private': 'warn',
 			'local/code-parameter-properties-must-have-explicit-accessibility': 'warn',
@@ -89,7 +89,9 @@ export default tseslint.config(
 			'local/code-declare-service-brand': 'warn',
 			'local/code-no-reader-after-await': 'warn',
 			'local/code-no-observable-get-in-reactive-context': 'warn',
+			'local/code-no-localized-model-description': 'warn',
 			'local/code-policy-localization-key-match': 'warn',
+			'local/code-no-localization-template-literals': 'error',
 			'local/code-no-deep-import-of-internal': ['error', { '.*Internal': true, 'searchExtTypesInternal': false }],
 			'local/code-layering': [
 				'warn',
@@ -131,7 +133,7 @@ export default tseslint.config(
 	// TS
 	{
 		files: [
-			'**/*.ts',
+			'**/*.{ts,tsx,mts,cts}',
 		],
 		languageOptions: {
 			parser: tseslint.parser,
@@ -143,6 +145,8 @@ export default tseslint.config(
 			'jsdoc': pluginJsdoc,
 		},
 		rules: {
+			// Disable built-in semi rules in favor of stylistic
+			'semi': 'off',
 			'@stylistic/ts/semi': 'warn',
 			'@stylistic/ts/member-delimiter-style': 'warn',
 			'local/code-no-unused-expressions': [
@@ -181,7 +185,8 @@ export default tseslint.config(
 	// Disallow 'in' operator except in type predicates
 	{
 		files: [
-			'**/*.ts'
+			'**/*.ts',
+			'.eslint-plugin-local/**/*.ts', // Explicitly include files under dot directories
 		],
 		ignores: [
 			'src/bootstrap-node.ts',
@@ -190,13 +195,7 @@ export default tseslint.config(
 			'extensions/debug-auto-launch/src/extension.ts',
 			'extensions/emmet/src/updateImageSize.ts',
 			'extensions/emmet/src/util.ts',
-			'extensions/git/src/blame.ts',
-			'extensions/github/src/links.ts',
 			'extensions/github-authentication/src/node/fetch.ts',
-			'extensions/ipynb/src/deserializers.ts',
-			'extensions/ipynb/src/notebookImagePaste.ts',
-			'extensions/ipynb/src/serializers.ts',
-			'extensions/notebook-renderers/src/index.ts',
 			'extensions/terminal-suggest/src/fig/figInterface.ts',
 			'extensions/terminal-suggest/src/fig/fig-autocomplete-shared/mixins.ts',
 			'extensions/terminal-suggest/src/fig/fig-autocomplete-shared/specMetadata.ts',
@@ -210,7 +209,6 @@ export default tseslint.config(
 			'src/vs/base/browser/dom.ts',
 			'src/vs/base/browser/markdownRenderer.ts',
 			'src/vs/base/browser/touch.ts',
-			'src/vs/base/browser/webWorkerFactory.ts',
 			'src/vs/base/common/async.ts',
 			'src/vs/base/common/desktopEnvironmentInfo.ts',
 			'src/vs/base/common/objects.ts',
@@ -232,7 +230,6 @@ export default tseslint.config(
 			'src/vs/editor/contrib/inlineCompletions/browser/model/provideInlineCompletions.ts',
 			'src/vs/editor/contrib/inlineCompletions/browser/view/ghostText/ghostTextView.ts',
 			'src/vs/editor/contrib/inlineCompletions/browser/view/inlineEdits/inlineEditsViews/debugVisualization.ts',
-			'src/vs/editor/contrib/quickAccess/browser/editorNavigationQuickAccess.ts',
 			'src/vs/platform/accessibilitySignal/browser/accessibilitySignalService.ts',
 			'src/vs/platform/configuration/common/configuration.ts',
 			'src/vs/platform/configuration/common/configurationModels.ts',
@@ -253,7 +250,6 @@ export default tseslint.config(
 			'src/vs/workbench/api/common/extHostNotebookKernels.ts',
 			'src/vs/workbench/api/common/extHostQuickOpen.ts',
 			'src/vs/workbench/api/common/extHostRequireInterceptor.ts',
-			'src/vs/workbench/api/common/extHostTelemetry.ts',
 			'src/vs/workbench/api/common/extHostTypeConverters.ts',
 			'src/vs/workbench/api/common/extHostTypes.ts',
 			'src/vs/workbench/api/node/loopbackServer.ts',
@@ -266,8 +262,6 @@ export default tseslint.config(
 			'src/vs/workbench/browser/workbench.ts',
 			'src/vs/workbench/common/notifications.ts',
 			'src/vs/workbench/contrib/accessibility/browser/accessibleView.ts',
-			'src/vs/workbench/contrib/chat/browser/actions/chatExecuteActions.ts',
-			'src/vs/workbench/contrib/chat/browser/chat.ts',
 			'src/vs/workbench/contrib/chat/browser/chatAttachmentResolveService.ts',
 			'src/vs/workbench/contrib/chat/browser/chatContentParts/chatAttachmentsContentPart.ts',
 			'src/vs/workbench/contrib/chat/browser/chatContentParts/chatConfirmationWidget.ts',
@@ -277,16 +271,10 @@ export default tseslint.config(
 			'src/vs/workbench/contrib/chat/browser/chatContentParts/toolInvocationParts/abstractToolConfirmationSubPart.ts',
 			'src/vs/workbench/contrib/chat/browser/chatEditing/chatEditingSession.ts',
 			'src/vs/workbench/contrib/chat/browser/chatEditing/chatEditingSessionStorage.ts',
-			'src/vs/workbench/contrib/chat/browser/chatEditorInput.ts',
-			'src/vs/workbench/contrib/chat/browser/chatFollowups.ts',
 			'src/vs/workbench/contrib/chat/browser/chatInlineAnchorWidget.ts',
-			'src/vs/workbench/contrib/chat/browser/chatInputPart.ts',
-			'src/vs/workbench/contrib/chat/browser/chatListRenderer.ts',
 			'src/vs/workbench/contrib/chat/browser/chatResponseAccessibleView.ts',
 			'src/vs/workbench/contrib/chat/browser/chatSessions/common.ts',
-			'src/vs/workbench/contrib/chat/browser/chatSessions/localChatSessionsProvider.ts',
 			'src/vs/workbench/contrib/chat/browser/chatSessions/view/sessionsTreeRenderer.ts',
-			'src/vs/workbench/contrib/chat/browser/chatWidget.ts',
 			'src/vs/workbench/contrib/chat/browser/contrib/chatInputCompletions.ts',
 			'src/vs/workbench/contrib/chat/common/annotations.ts',
 			'src/vs/workbench/contrib/chat/common/chat.ts',
@@ -321,58 +309,19 @@ export default tseslint.config(
 			'src/vs/workbench/contrib/mcp/common/mcpServerRequestHandler.ts',
 			'src/vs/workbench/contrib/mcp/test/common/mcpRegistryTypes.ts',
 			'src/vs/workbench/contrib/mcp/test/common/mcpServerRequestHandler.test.ts',
-			'src/vs/workbench/contrib/notebook/browser/contrib/debug/notebookBreakpoints.ts',
-			'src/vs/workbench/contrib/notebook/browser/contrib/find/findModel.ts',
 			'src/vs/workbench/contrib/notebook/browser/controller/cellOutputActions.ts',
 			'src/vs/workbench/contrib/notebook/browser/controller/chat/notebook.chat.contribution.ts',
 			'src/vs/workbench/contrib/notebook/browser/controller/coreActions.ts',
-			'src/vs/workbench/contrib/notebook/browser/diff/diffElementViewModel.ts',
-			'src/vs/workbench/contrib/notebook/browser/view/cellParts/cellExecution.ts',
 			'src/vs/workbench/contrib/notebook/browser/view/renderers/backLayerWebView.ts',
 			'src/vs/workbench/contrib/notebook/browser/viewParts/notebookKernelView.ts',
-			'src/vs/workbench/contrib/notebook/common/model/notebookTextModel.ts',
-			'src/vs/workbench/contrib/notebook/common/notebookEditorModel.ts',
 			'src/vs/workbench/contrib/output/browser/outputView.ts',
 			'src/vs/workbench/contrib/preferences/browser/settingsTree.ts',
 			'src/vs/workbench/contrib/remoteTunnel/electron-browser/remoteTunnel.contribution.ts',
 			'src/vs/workbench/contrib/tasks/browser/abstractTaskService.ts',
 			'src/vs/workbench/contrib/tasks/browser/taskTerminalStatus.ts',
 			'src/vs/workbench/contrib/tasks/browser/terminalTaskSystem.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/baseTerminalBackend.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/remotePty.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/remoteTerminalBackend.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalActions.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalEditorService.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalGroup.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalIcon.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalIconPicker.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalInstance.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalInstanceService.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalMenus.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalProfileQuickpick.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalProfileResolverService.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalProfileService.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalService.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalTabsList.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/terminalView.ts',
-			// 'src/vs/workbench/contrib/terminal/browser/xterm/markNavigationAddon.ts',
-			'src/vs/workbench/contrib/terminalContrib/accessibility/browser/terminal.accessibility.contribution.ts',
-			'src/vs/workbench/contrib/terminalContrib/accessibility/browser/terminalAccessibleBufferProvider.ts',
-			'src/vs/workbench/contrib/terminalContrib/chat/browser/terminal.initialHint.contribution.ts',
-			'src/vs/workbench/contrib/terminalContrib/chat/browser/terminalChatActions.ts',
-			'src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/commandLineAutoApprover.ts',
 			'src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/taskHelpers.ts',
-			'src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/toolTerminalCreator.ts',
 			'src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/tools/monitoring/outputMonitor.ts',
-			'src/vs/workbench/contrib/terminalContrib/commandGuide/browser/terminal.commandGuide.contribution.ts',
-			'src/vs/workbench/contrib/terminalContrib/history/browser/terminalRunRecentQuickPick.ts',
-			'src/vs/workbench/contrib/terminalContrib/links/browser/terminalLinkQuickpick.ts',
-			'src/vs/workbench/contrib/terminalContrib/links/test/browser/linkTestUtils.ts',
-			'src/vs/workbench/contrib/terminalContrib/quickFix/browser/quickFixAddon.ts',
-			'src/vs/workbench/contrib/terminalContrib/sendSequence/browser/terminal.sendSequence.contribution.ts',
-			'src/vs/workbench/contrib/terminalContrib/sendSignal/browser/terminal.sendSignal.contribution.ts',
-			'src/vs/workbench/contrib/terminalContrib/stickyScroll/browser/terminalStickyScrollOverlay.ts',
-			'src/vs/workbench/contrib/terminalContrib/suggest/browser/terminalCompletionService.ts',
 			'src/vs/workbench/contrib/testing/browser/explorerProjections/listProjection.ts',
 			'src/vs/workbench/contrib/testing/browser/explorerProjections/treeProjection.ts',
 			'src/vs/workbench/contrib/testing/browser/testCoverageBars.ts',
@@ -391,8 +340,6 @@ export default tseslint.config(
 			'src/vs/workbench/services/preferences/common/preferencesValidation.ts',
 			'src/vs/workbench/services/remote/common/tunnelModel.ts',
 			'src/vs/workbench/services/search/common/textSearchManager.ts',
-			'src/vs/workbench/services/telemetry/test/browser/commonProperties.test.ts',
-			'src/vs/workbench/services/telemetry/test/node/commonProperties.test.ts',
 			'src/vs/workbench/test/browser/workbenchTestServices.ts',
 			'test/automation/src/playwrightDriver.ts',
 			'.eslint-plugin-local/**/*',
@@ -404,62 +351,55 @@ export default tseslint.config(
 			'local/code-no-in-operator': 'warn',
 		}
 	},
-	// vscode TS: strict no explicit `any`
+	// Strict no explicit `any`
 	{
 		files: [
+			// Extensions
+			'extensions/git/src/**/*.ts',
+			'extensions/git-base/src/**/*.ts',
+			'extensions/github/src/**/*.ts',
+			// vscode
 			'src/**/*.ts',
 		],
 		ignores: [
+			// Extensions
+			'extensions/git/src/commands.ts',
+			'extensions/git/src/decorators.ts',
+			'extensions/git/src/git.ts',
+			'extensions/git/src/util.ts',
+			'extensions/git-base/src/decorators.ts',
+			'extensions/github/src/util.ts',
+			// vscode d.ts
 			'src/vs/amdX.ts',
 			'src/vs/monaco.d.ts',
 			'src/vscode-dts/**',
 			// Base
 			'src/vs/base/browser/dom.ts',
 			'src/vs/base/browser/mouseEvent.ts',
-			'src/vs/base/browser/pixelRatio.ts',
-			'src/vs/base/browser/trustedTypes.ts',
-			'src/vs/base/browser/webWorkerFactory.ts',
-			'src/vs/base/node/osDisplayProtocolInfo.ts',
-			'src/vs/base/node/osReleaseInfo.ts',
 			'src/vs/base/node/processes.ts',
 			'src/vs/base/common/arrays.ts',
 			'src/vs/base/common/async.ts',
-			'src/vs/base/common/cancellation.ts',
-			'src/vs/base/common/collections.ts',
 			'src/vs/base/common/console.ts',
-			'src/vs/base/common/controlFlow.ts',
 			'src/vs/base/common/decorators.ts',
-			'src/vs/base/common/equals.ts',
 			'src/vs/base/common/errorMessage.ts',
 			'src/vs/base/common/errors.ts',
 			'src/vs/base/common/event.ts',
 			'src/vs/base/common/hotReload.ts',
 			'src/vs/base/common/hotReloadHelpers.ts',
-			'src/vs/base/common/iterator.ts',
 			'src/vs/base/common/json.ts',
 			'src/vs/base/common/jsonSchema.ts',
 			'src/vs/base/common/lifecycle.ts',
-			'src/vs/base/common/linkedList.ts',
 			'src/vs/base/common/map.ts',
 			'src/vs/base/common/marshalling.ts',
-			'src/vs/base/common/network.ts',
-			'src/vs/base/common/oauth.ts',
 			'src/vs/base/common/objects.ts',
 			'src/vs/base/common/performance.ts',
 			'src/vs/base/common/platform.ts',
 			'src/vs/base/common/processes.ts',
-			'src/vs/base/common/resourceTree.ts',
-			'src/vs/base/common/skipList.ts',
-			'src/vs/base/common/strings.ts',
-			'src/vs/base/common/ternarySearchTree.ts',
 			'src/vs/base/common/types.ts',
 			'src/vs/base/common/uriIpc.ts',
 			'src/vs/base/common/verifier.ts',
 			'src/vs/base/common/observableInternal/base.ts',
 			'src/vs/base/common/observableInternal/changeTracker.ts',
-			'src/vs/base/common/observableInternal/debugLocation.ts',
-			'src/vs/base/common/observableInternal/debugName.ts',
-			'src/vs/base/common/observableInternal/map.ts',
 			'src/vs/base/common/observableInternal/set.ts',
 			'src/vs/base/common/observableInternal/transaction.ts',
 			'src/vs/base/common/worker/webWorkerBootstrap.ts',
@@ -477,15 +417,6 @@ export default tseslint.config(
 			'src/vs/base/browser/ui/list/rowCache.ts',
 			'src/vs/base/browser/ui/sash/sash.ts',
 			'src/vs/base/browser/ui/table/tableWidget.ts',
-			'src/vs/base/browser/ui/tree/abstractTree.ts',
-			'src/vs/base/browser/ui/tree/asyncDataTree.ts',
-			'src/vs/base/browser/ui/tree/compressedObjectTreeModel.ts',
-			'src/vs/base/browser/ui/tree/dataTree.ts',
-			'src/vs/base/browser/ui/tree/indexTree.ts',
-			'src/vs/base/browser/ui/tree/indexTreeModel.ts',
-			'src/vs/base/browser/ui/tree/objectTree.ts',
-			'src/vs/base/browser/ui/tree/objectTreeModel.ts',
-			'src/vs/base/browser/ui/tree/tree.ts',
 			'src/vs/base/parts/ipc/common/ipc.net.ts',
 			'src/vs/base/parts/ipc/common/ipc.ts',
 			'src/vs/base/parts/ipc/electron-main/ipcMain.ts',
@@ -509,46 +440,15 @@ export default tseslint.config(
 			'src/vs/base/common/observableInternal/logging/debugger/rpc.ts',
 			'src/vs/base/test/browser/ui/grid/util.ts',
 			// Platform
-			'src/vs/platform/accessibility/browser/accessibleView.ts',
-			'src/vs/platform/accessibility/common/accessibility.ts',
-			'src/vs/platform/action/common/action.ts',
-			'src/vs/platform/actions/common/actions.ts',
-			'src/vs/platform/assignment/common/assignment.ts',
-			'src/vs/platform/browserElements/electron-main/nativeBrowserElementsMainService.ts',
 			'src/vs/platform/commands/common/commands.ts',
-			'src/vs/platform/configuration/common/configuration.ts',
-			'src/vs/platform/configuration/common/configurationModels.ts',
-			'src/vs/platform/configuration/common/configurationRegistry.ts',
-			'src/vs/platform/configuration/common/configurationService.ts',
-			'src/vs/platform/configuration/common/configurations.ts',
 			'src/vs/platform/contextkey/browser/contextKeyService.ts',
 			'src/vs/platform/contextkey/common/contextkey.ts',
 			'src/vs/platform/contextview/browser/contextView.ts',
-			'src/vs/platform/contextview/browser/contextViewService.ts',
 			'src/vs/platform/debug/common/extensionHostDebugIpc.ts',
 			'src/vs/platform/debug/electron-main/extensionHostDebugIpc.ts',
 			'src/vs/platform/diagnostics/common/diagnostics.ts',
-			'src/vs/platform/diagnostics/node/diagnosticsService.ts',
 			'src/vs/platform/download/common/downloadIpc.ts',
-			'src/vs/platform/extensionManagement/common/abstractExtensionManagementService.ts',
-			'src/vs/platform/extensionManagement/common/allowedExtensionsService.ts',
-			'src/vs/platform/extensionManagement/common/extensionGalleryManifestServiceIpc.ts',
-			'src/vs/platform/extensionManagement/common/extensionGalleryService.ts',
-			'src/vs/platform/extensionManagement/common/extensionManagement.ts',
-			'src/vs/platform/extensionManagement/common/extensionManagementIpc.ts',
-			'src/vs/platform/extensionManagement/common/extensionManagementUtil.ts',
-			'src/vs/platform/extensionManagement/common/extensionNls.ts',
-			'src/vs/platform/extensionManagement/common/extensionStorage.ts',
-			'src/vs/platform/extensionManagement/common/extensionTipsService.ts',
-			'src/vs/platform/extensionManagement/common/extensionsProfileScannerService.ts',
-			'src/vs/platform/extensionManagement/common/implicitActivationEvents.ts',
-			'src/vs/platform/extensionManagement/node/extensionManagementService.ts',
-			'src/vs/platform/extensionRecommendations/common/extensionRecommendationsIpc.ts',
-			'src/vs/platform/extensions/common/extensionHostStarter.ts',
-			'src/vs/platform/extensions/common/extensionValidator.ts',
 			'src/vs/platform/extensions/common/extensions.ts',
-			'src/vs/platform/extensions/electron-main/extensionHostStarter.ts',
-			'src/vs/platform/externalTerminal/node/externalTerminalService.ts',
 			'src/vs/platform/instantiation/common/descriptors.ts',
 			'src/vs/platform/instantiation/common/extensions.ts',
 			'src/vs/platform/instantiation/common/instantiation.ts',
@@ -558,26 +458,21 @@ export default tseslint.config(
 			'src/vs/platform/keybinding/common/keybindingResolver.ts',
 			'src/vs/platform/keybinding/common/keybindingsRegistry.ts',
 			'src/vs/platform/keybinding/common/resolvedKeybindingItem.ts',
-			'src/vs/platform/keyboardLayout/common/keyboardConfig.ts',
 			'src/vs/platform/languagePacks/node/languagePacks.ts',
 			'src/vs/platform/list/browser/listService.ts',
 			'src/vs/platform/log/browser/log.ts',
 			'src/vs/platform/log/common/log.ts',
 			'src/vs/platform/log/common/logIpc.ts',
-			'src/vs/workbench/disablechat.ts',
 			'src/vs/platform/log/electron-main/logIpc.ts',
 			'src/vs/platform/observable/common/wrapInHotClass.ts',
 			'src/vs/platform/observable/common/wrapInReloadableClass.ts',
 			'src/vs/platform/policy/common/policyIpc.ts',
-			'src/vs/platform/policy/node/nativePolicyService.ts',
 			'src/vs/platform/profiling/common/profilingTelemetrySpec.ts',
 			'src/vs/platform/quickinput/browser/quickInputActions.ts',
 			'src/vs/platform/quickinput/common/quickInput.ts',
 			'src/vs/platform/registry/common/platform.ts',
 			'src/vs/platform/remote/browser/browserSocketFactory.ts',
 			'src/vs/platform/remote/browser/remoteAuthorityResolverService.ts',
-			'src/vs/platform/remote/common/electronRemoteResources.ts',
-			'src/vs/platform/remote/common/managedSocket.ts',
 			'src/vs/platform/remote/common/remoteAgentConnection.ts',
 			'src/vs/platform/remote/common/remoteAuthorityResolver.ts',
 			'src/vs/platform/remote/electron-browser/electronRemoteResourceLoader.ts',
@@ -587,24 +482,10 @@ export default tseslint.config(
 			'src/vs/platform/request/common/requestIpc.ts',
 			'src/vs/platform/request/electron-utility/requestService.ts',
 			'src/vs/platform/request/node/proxy.ts',
-			'src/vs/platform/telemetry/browser/1dsAppender.ts',
 			'src/vs/platform/telemetry/browser/errorTelemetry.ts',
-			'src/vs/platform/telemetry/common/1dsAppender.ts',
 			'src/vs/platform/telemetry/common/errorTelemetry.ts',
-			'src/vs/platform/telemetry/common/gdprTypings.ts',
 			'src/vs/platform/telemetry/common/remoteTelemetryChannel.ts',
-			'src/vs/platform/telemetry/common/telemetry.ts',
-			'src/vs/platform/telemetry/common/telemetryIpc.ts',
-			'src/vs/platform/telemetry/common/telemetryLogAppender.ts',
-			'src/vs/platform/telemetry/common/telemetryService.ts',
-			'src/vs/platform/telemetry/common/telemetryUtils.ts',
-			'src/vs/platform/telemetry/node/1dsAppender.ts',
 			'src/vs/platform/telemetry/node/errorTelemetry.ts',
-			'src/vs/platform/terminal/common/terminal.ts',
-			'src/vs/platform/terminal/node/ptyHostService.ts',
-			'src/vs/platform/terminal/node/ptyService.ts',
-			'src/vs/platform/terminal/node/terminalProcess.ts',
-			'src/vs/platform/terminal/node/windowsShellHelper.ts',
 			'src/vs/platform/theme/common/iconRegistry.ts',
 			'src/vs/platform/theme/common/tokenClassificationRegistry.ts',
 			'src/vs/platform/update/common/updateIpc.ts',
@@ -617,26 +498,14 @@ export default tseslint.config(
 			'src/vs/platform/userDataSync/common/extensionsSync.ts',
 			'src/vs/platform/userDataSync/common/globalStateMerge.ts',
 			'src/vs/platform/userDataSync/common/globalStateSync.ts',
-			'src/vs/platform/userDataSync/common/ignoredExtensions.ts',
 			'src/vs/platform/userDataSync/common/settingsMerge.ts',
 			'src/vs/platform/userDataSync/common/settingsSync.ts',
-			'src/vs/platform/userDataSync/common/userDataAutoSyncService.ts',
 			'src/vs/platform/userDataSync/common/userDataSync.ts',
-			'src/vs/platform/userDataSync/common/userDataSyncAccount.ts',
-			'src/vs/platform/userDataSync/common/userDataSyncEnablementService.ts',
 			'src/vs/platform/userDataSync/common/userDataSyncIpc.ts',
-			'src/vs/platform/userDataSync/common/userDataSyncLocalStoreService.ts',
-			'src/vs/platform/userDataSync/common/userDataSyncMachines.ts',
-			'src/vs/platform/userDataSync/common/userDataSyncResourceProvider.ts',
-			'src/vs/platform/userDataSync/common/userDataSyncService.ts',
 			'src/vs/platform/userDataSync/common/userDataSyncServiceIpc.ts',
-			'src/vs/platform/userDataSync/common/userDataSyncStoreService.ts',
-			'src/vs/platform/webContentExtractor/electron-main/cdpAccessibilityDomain.ts',
 			'src/vs/platform/webview/common/webviewManagerService.ts',
-			'src/vs/platform/configuration/test/common/testConfigurationService.ts',
 			'src/vs/platform/instantiation/test/common/instantiationServiceMock.ts',
 			'src/vs/platform/keybinding/test/common/mockKeybindingService.ts',
-			'src/vs/platform/userDataSync/test/common/userDataSyncClient.ts',
 			// Editor
 			'src/vs/editor/standalone/browser/standaloneEditor.ts',
 			'src/vs/editor/standalone/browser/standaloneLanguages.ts',
@@ -647,16 +516,11 @@ export default tseslint.config(
 			'src/vs/editor/contrib/codeAction/browser/codeAction.ts',
 			'src/vs/editor/contrib/codeAction/browser/codeActionCommands.ts',
 			'src/vs/editor/contrib/codeAction/common/types.ts',
-			'src/vs/editor/contrib/codelens/browser/codelens.ts',
-			'src/vs/editor/contrib/codelens/browser/codelensController.ts',
 			'src/vs/editor/contrib/colorPicker/browser/colorDetector.ts',
 			'src/vs/editor/contrib/diffEditorBreadcrumbs/browser/contribution.ts',
 			'src/vs/editor/contrib/dropOrPasteInto/browser/dropIntoEditorContribution.ts',
 			'src/vs/editor/contrib/find/browser/findController.ts',
 			'src/vs/editor/contrib/find/browser/findModel.ts',
-			'src/vs/editor/contrib/find/browser/findWidgetSearchHistory.ts',
-			'src/vs/editor/contrib/find/browser/replaceWidgetHistory.ts',
-			'src/vs/editor/contrib/folding/browser/folding.ts',
 			'src/vs/editor/contrib/gotoSymbol/browser/goToCommands.ts',
 			'src/vs/editor/contrib/gotoSymbol/browser/symbolNavigation.ts',
 			'src/vs/editor/contrib/hover/browser/hoverActions.ts',
@@ -680,19 +544,14 @@ export default tseslint.config(
 			'src/vs/editor/contrib/inlineCompletions/browser/view/inlineEdits/utils/utils.ts',
 			// Workbench
 			'src/vs/workbench/api/browser/mainThreadChatSessions.ts',
-			'src/vs/workbench/api/browser/mainThreadTerminalService.ts',
-			'src/vs/workbench/api/common/configurationExtensionPoint.ts',
 			'src/vs/workbench/api/common/extHost.api.impl.ts',
 			'src/vs/workbench/api/common/extHost.protocol.ts',
 			'src/vs/workbench/api/common/extHostChatSessions.ts',
 			'src/vs/workbench/api/common/extHostCodeInsets.ts',
 			'src/vs/workbench/api/common/extHostCommands.ts',
-			'src/vs/workbench/api/common/extHostConfiguration.ts',
 			'src/vs/workbench/api/common/extHostConsoleForwarder.ts',
 			'src/vs/workbench/api/common/extHostDataChannels.ts',
 			'src/vs/workbench/api/common/extHostDebugService.ts',
-			'src/vs/workbench/api/common/extHostDiagnostics.ts',
-			'src/vs/workbench/api/common/extHostDocumentSaveParticipant.ts',
 			'src/vs/workbench/api/common/extHostExtensionActivator.ts',
 			'src/vs/workbench/api/common/extHostExtensionService.ts',
 			'src/vs/workbench/api/common/extHostFileSystemConsumer.ts',
@@ -704,7 +563,6 @@ export default tseslint.config(
 			'src/vs/workbench/api/common/extHostMessageService.ts',
 			'src/vs/workbench/api/common/extHostNotebookDocument.ts',
 			'src/vs/workbench/api/common/extHostNotebookDocumentSaveParticipant.ts',
-			'src/vs/workbench/api/common/extHostQuickOpen.ts',
 			'src/vs/workbench/api/common/extHostRequireInterceptor.ts',
 			'src/vs/workbench/api/common/extHostRpcService.ts',
 			'src/vs/workbench/api/common/extHostSCM.ts',
@@ -712,7 +570,6 @@ export default tseslint.config(
 			'src/vs/workbench/api/common/extHostStatusBar.ts',
 			'src/vs/workbench/api/common/extHostStoragePaths.ts',
 			'src/vs/workbench/api/common/extHostTelemetry.ts',
-			'src/vs/workbench/api/common/extHostTerminalService.ts',
 			'src/vs/workbench/api/common/extHostTesting.ts',
 			'src/vs/workbench/api/common/extHostTextEditor.ts',
 			'src/vs/workbench/api/common/extHostTimeline.ts',
@@ -743,48 +600,13 @@ export default tseslint.config(
 			'src/vs/workbench/contrib/accessibility/browser/accessibilityConfiguration.ts',
 			'src/vs/workbench/contrib/accessibilitySignals/browser/commands.ts',
 			'src/vs/workbench/contrib/authentication/browser/actions/manageTrustedMcpServersForAccountAction.ts',
-			'src/vs/workbench/contrib/bulkEdit/browser/bulkCellEdits.ts',
 			'src/vs/workbench/contrib/bulkEdit/browser/bulkTextEdits.ts',
-			'src/vs/workbench/contrib/bulkEdit/browser/opaqueEdits.ts',
 			'src/vs/workbench/contrib/bulkEdit/browser/preview/bulkEditPane.ts',
 			'src/vs/workbench/contrib/bulkEdit/browser/preview/bulkEditPreview.ts',
-			'src/vs/workbench/contrib/callHierarchy/browser/callHierarchy.contribution.ts',
-			'src/vs/workbench/contrib/callHierarchy/common/callHierarchy.ts',
-			'src/vs/workbench/contrib/chat/browser/actions/chatCodeblockActions.ts',
-			'src/vs/workbench/contrib/chat/browser/actions/chatContextActions.ts',
-			'src/vs/workbench/contrib/chat/browser/actions/chatToolActions.ts',
-			'src/vs/workbench/contrib/chat/browser/chatAttachmentWidgets.ts',
-			'src/vs/workbench/contrib/chat/browser/chatContentParts/chatConfirmationWidget.ts',
-			'src/vs/workbench/contrib/chat/browser/chatContentParts/chatMultiDiffContentPart.ts',
-			'src/vs/workbench/contrib/chat/browser/chatEditing/chatEditingActions.ts',
-			'src/vs/workbench/contrib/chat/browser/chatEditing/chatEditingEditorActions.ts',
-			'src/vs/workbench/contrib/chat/browser/chatEditing/chatEditingServiceImpl.ts',
-			'src/vs/workbench/contrib/chat/browser/chatInputPart.ts',
-			'src/vs/workbench/contrib/chat/browser/chatSessions.contribution.ts',
-			'src/vs/workbench/contrib/chat/browser/chatSessions/common.ts',
-			'src/vs/workbench/contrib/chat/browser/chatSessions/view/sessionsTreeRenderer.ts',
-			'src/vs/workbench/contrib/chat/browser/chatWidget.ts',
-			'src/vs/workbench/contrib/chat/browser/contrib/chatDynamicVariables.ts',
-			'src/vs/workbench/contrib/chat/common/chatAgents.ts',
-			'src/vs/workbench/contrib/chat/common/chatModel.ts',
-			'src/vs/workbench/contrib/chat/common/chatModes.ts',
-			'src/vs/workbench/contrib/chat/common/chatService.ts',
-			'src/vs/workbench/contrib/chat/common/chatServiceImpl.ts',
-			'src/vs/workbench/contrib/chat/common/chatSessionsService.ts',
-			'src/vs/workbench/contrib/chat/common/chatWidgetHistoryService.ts',
-			'src/vs/workbench/contrib/chat/common/languageModelToolsService.ts',
-			'src/vs/workbench/contrib/chat/common/languageModels.ts',
-			'src/vs/workbench/contrib/chat/common/promptSyntax/service/promptsServiceImpl.ts',
-			'src/vs/workbench/contrib/chat/common/tools/manageTodoListTool.ts',
-			'src/vs/workbench/contrib/chat/test/common/languageModels.ts',
-			'src/vs/workbench/contrib/chat/test/common/mockLanguageModelToolsService.ts',
-			'src/vs/workbench/contrib/chat/test/common/mockPromptsService.ts',
 			'src/vs/workbench/contrib/codeEditor/browser/inspectEditorTokens/inspectEditorTokens.ts',
 			'src/vs/workbench/contrib/codeEditor/browser/outline/documentSymbolsOutline.ts',
 			'src/vs/workbench/contrib/codeEditor/electron-browser/selectionClipboard.ts',
 			'src/vs/workbench/contrib/commands/common/commands.contribution.ts',
-			'src/vs/workbench/contrib/comments/browser/commentNode.ts',
-			'src/vs/workbench/contrib/comments/browser/commentsAccessibleView.ts',
 			'src/vs/workbench/contrib/comments/browser/commentsTreeViewer.ts',
 			'src/vs/workbench/contrib/comments/browser/commentsView.ts',
 			'src/vs/workbench/contrib/comments/browser/reactionsAction.ts',
@@ -809,7 +631,6 @@ export default tseslint.config(
 			'src/vs/workbench/contrib/debug/common/debugger.ts',
 			'src/vs/workbench/contrib/debug/common/replModel.ts',
 			'src/vs/workbench/contrib/debug/test/common/mockDebug.ts',
-			'src/vs/workbench/contrib/editSessions/common/editSessionsStorageClient.ts',
 			'src/vs/workbench/contrib/editSessions/common/workspaceStateSync.ts',
 			'src/vs/workbench/contrib/editTelemetry/browser/helpers/documentWithAnnotatedEdits.ts',
 			'src/vs/workbench/contrib/editTelemetry/browser/helpers/utils.ts',
@@ -823,18 +644,12 @@ export default tseslint.config(
 			'src/vs/workbench/contrib/extensions/browser/extensionsViews.ts',
 			'src/vs/workbench/contrib/extensions/browser/extensionsWorkbenchService.ts',
 			'src/vs/workbench/contrib/extensions/common/extensions.ts',
-			'src/vs/workbench/contrib/extensions/electron-browser/extensionProfileService.ts',
 			'src/vs/workbench/contrib/extensions/electron-browser/runtimeExtensionsEditor.ts',
-			'src/vs/workbench/contrib/format/browser/formatActionsMultiple.ts',
 			'src/vs/workbench/contrib/inlineChat/browser/inlineChatActions.ts',
 			'src/vs/workbench/contrib/inlineChat/browser/inlineChatController.ts',
 			'src/vs/workbench/contrib/inlineChat/browser/inlineChatStrategies.ts',
-			'src/vs/workbench/contrib/issue/browser/issueReporterModel.ts',
-			'src/vs/workbench/contrib/list/browser/tableColumnResizeQuickPick.ts',
 			'src/vs/workbench/contrib/markdown/browser/markdownDocumentRenderer.ts',
-			'src/vs/workbench/contrib/markdown/browser/markdownSettingRenderer.ts',
 			'src/vs/workbench/contrib/markers/browser/markers.contribution.ts',
-			'src/vs/workbench/contrib/markers/browser/markersTable.ts',
 			'src/vs/workbench/contrib/markers/browser/markersView.ts',
 			'src/vs/workbench/contrib/mergeEditor/browser/commands/commands.ts',
 			'src/vs/workbench/contrib/mergeEditor/browser/utils.ts',
@@ -843,7 +658,6 @@ export default tseslint.config(
 			'src/vs/workbench/contrib/notebook/browser/contrib/clipboard/notebookClipboard.ts',
 			'src/vs/workbench/contrib/notebook/browser/contrib/find/notebookFind.ts',
 			'src/vs/workbench/contrib/notebook/browser/contrib/layout/layoutActions.ts',
-			'src/vs/workbench/contrib/notebook/browser/contrib/multicursor/notebookMulticursor.ts',
 			'src/vs/workbench/contrib/notebook/browser/contrib/profile/notebookProfile.ts',
 			'src/vs/workbench/contrib/notebook/browser/contrib/troubleshoot/layout.ts',
 			'src/vs/workbench/contrib/notebook/browser/controller/chat/cellChatActions.ts',
@@ -854,29 +668,23 @@ export default tseslint.config(
 			'src/vs/workbench/contrib/notebook/browser/diff/diffComponents.ts',
 			'src/vs/workbench/contrib/notebook/browser/diff/inlineDiff/notebookDeletedCellDecorator.ts',
 			'src/vs/workbench/contrib/notebook/browser/notebookBrowser.ts',
-			'src/vs/workbench/contrib/notebook/browser/notebookEditorWidget.ts',
 			'src/vs/workbench/contrib/notebook/browser/outputEditor/notebookOutputEditor.ts',
 			'src/vs/workbench/contrib/notebook/browser/services/notebookEditorServiceImpl.ts',
 			'src/vs/workbench/contrib/notebook/browser/view/notebookCellList.ts',
 			'src/vs/workbench/contrib/notebook/browser/view/renderers/backLayerWebView.ts',
 			'src/vs/workbench/contrib/notebook/browser/view/renderers/webviewMessages.ts',
 			'src/vs/workbench/contrib/notebook/browser/view/renderers/webviewPreloads.ts',
-			'src/vs/workbench/contrib/notebook/browser/viewModel/cellEditorOptions.ts',
 			'src/vs/workbench/contrib/notebook/browser/viewModel/markupCellViewModel.ts',
 			'src/vs/workbench/contrib/notebook/browser/viewParts/notebookEditorStickyScroll.ts',
 			'src/vs/workbench/contrib/notebook/browser/viewParts/notebookHorizontalTracker.ts',
 			'src/vs/workbench/contrib/notebook/browser/viewParts/notebookKernelQuickPickStrategy.ts',
-			'src/vs/workbench/contrib/notebook/browser/viewParts/notebookViewZones.ts',
-			'src/vs/workbench/contrib/notebook/common/model/notebookCellOutputTextModel.ts',
 			'src/vs/workbench/contrib/notebook/common/model/notebookCellTextModel.ts',
 			'src/vs/workbench/contrib/notebook/common/model/notebookMetadataTextModel.ts',
 			'src/vs/workbench/contrib/notebook/common/model/notebookTextModel.ts',
 			'src/vs/workbench/contrib/notebook/common/notebookCommon.ts',
 			'src/vs/workbench/contrib/notebook/common/notebookEditorModelResolverServiceImpl.ts',
-			'src/vs/workbench/contrib/notebook/common/notebookRange.ts',
 			'src/vs/workbench/contrib/notebook/test/browser/testNotebookEditor.ts',
 			'src/vs/workbench/contrib/performance/electron-browser/startupProfiler.ts',
-			'src/vs/workbench/contrib/preferences/browser/keybindingsEditor.ts',
 			'src/vs/workbench/contrib/preferences/browser/preferences.contribution.ts',
 			'src/vs/workbench/contrib/preferences/browser/preferencesRenderers.ts',
 			'src/vs/workbench/contrib/preferences/browser/settingsEditor2.ts',
@@ -905,15 +713,12 @@ export default tseslint.config(
 			'src/vs/workbench/contrib/search/browser/searchTreeModel/searchTreeCommon.ts',
 			'src/vs/workbench/contrib/search/browser/searchTreeModel/textSearchHeading.ts',
 			'src/vs/workbench/contrib/search/browser/searchView.ts',
-			'src/vs/workbench/contrib/search/browser/searchWidget.ts',
-			'src/vs/workbench/contrib/search/common/cacheState.ts',
 			'src/vs/workbench/contrib/search/test/browser/mockSearchTree.ts',
 			'src/vs/workbench/contrib/searchEditor/browser/searchEditor.contribution.ts',
 			'src/vs/workbench/contrib/searchEditor/browser/searchEditorActions.ts',
 			'src/vs/workbench/contrib/searchEditor/browser/searchEditorInput.ts',
 			'src/vs/workbench/contrib/snippets/browser/commands/configureSnippets.ts',
 			'src/vs/workbench/contrib/snippets/browser/commands/insertSnippet.ts',
-			'src/vs/workbench/contrib/snippets/browser/snippetsFile.ts',
 			'src/vs/workbench/contrib/snippets/browser/snippetsService.ts',
 			'src/vs/workbench/contrib/tasks/browser/abstractTaskService.ts',
 			'src/vs/workbench/contrib/tasks/browser/runAutomaticTasks.ts',
@@ -925,110 +730,54 @@ export default tseslint.config(
 			'src/vs/workbench/contrib/tasks/common/taskConfiguration.ts',
 			'src/vs/workbench/contrib/tasks/common/taskSystem.ts',
 			'src/vs/workbench/contrib/tasks/common/tasks.ts',
-			'src/vs/workbench/contrib/terminal/browser/remoteTerminalBackend.ts',
-			'src/vs/workbench/contrib/terminal/browser/terminalConfigurationService.ts',
-			'src/vs/workbench/contrib/terminal/browser/terminalExtensions.ts',
-			'src/vs/workbench/contrib/terminal/browser/terminalProcessExtHostProxy.ts',
-			'src/vs/workbench/contrib/terminal/browser/terminalProcessManager.ts',
-			'src/vs/workbench/contrib/terminal/browser/terminalProfileQuickpick.ts',
-			'src/vs/workbench/contrib/terminal/browser/terminalProfileService.ts',
-			'src/vs/workbench/contrib/terminal/browser/xterm/xtermTerminal.ts',
-			'src/vs/workbench/contrib/terminal/common/basePty.ts',
-			'src/vs/workbench/contrib/terminal/common/remote/remoteTerminalChannel.ts',
-			'src/vs/workbench/contrib/terminal/common/terminal.ts',
-			'src/vs/workbench/contrib/terminalContrib/links/browser/links.ts',
-			'src/vs/workbench/contrib/terminalContrib/suggest/browser/terminalSuggestAddon.ts',
-			'src/vs/workbench/contrib/testing/browser/testExplorerActions.ts',
-			'src/vs/workbench/contrib/testing/browser/testingExplorerView.ts',
 			'src/vs/workbench/contrib/testing/common/storedValue.ts',
-			'src/vs/workbench/contrib/testing/common/testItemCollection.ts',
 			'src/vs/workbench/contrib/testing/test/browser/testObjectTree.ts',
-			'src/vs/workbench/contrib/timeline/browser/timelinePane.ts',
 			'src/vs/workbench/contrib/typeHierarchy/browser/typeHierarchy.contribution.ts',
 			'src/vs/workbench/contrib/typeHierarchy/common/typeHierarchy.ts',
-			'src/vs/workbench/contrib/update/browser/update.ts',
-			'src/vs/workbench/contrib/userDataSync/browser/userDataSync.ts',
 			'src/vs/workbench/contrib/webview/browser/overlayWebview.ts',
 			'src/vs/workbench/contrib/webview/browser/webview.ts',
 			'src/vs/workbench/contrib/webview/browser/webviewElement.ts',
 			'src/vs/workbench/contrib/webviewPanel/browser/webviewEditor.ts',
 			'src/vs/workbench/contrib/webviewPanel/browser/webviewEditorInputSerializer.ts',
 			'src/vs/workbench/contrib/webviewPanel/browser/webviewWorkbenchService.ts',
-			'src/vs/workbench/contrib/webviewView/browser/webviewViewPane.ts',
-			'src/vs/workbench/contrib/welcomeGettingStarted/browser/gettingStarted.ts',
-			'src/vs/workbench/contrib/welcomeGettingStarted/browser/gettingStartedAccessibleView.ts',
 			'src/vs/workbench/contrib/welcomeGettingStarted/browser/gettingStartedService.ts',
-			'src/vs/workbench/contrib/welcomeViews/common/newFile.contribution.ts',
 			'src/vs/workbench/contrib/welcomeWalkthrough/browser/walkThroughPart.ts',
-			'src/vs/workbench/services/accounts/common/defaultAccount.ts',
-			'src/vs/workbench/services/assignment/common/assignmentFilters.ts',
 			'src/vs/workbench/services/authentication/common/authentication.ts',
 			'src/vs/workbench/services/authentication/test/browser/authenticationQueryServiceMocks.ts',
 			'src/vs/workbench/services/commands/common/commandService.ts',
-			'src/vs/workbench/services/configuration/browser/configuration.ts',
-			'src/vs/workbench/services/configuration/browser/configurationService.ts',
-			'src/vs/workbench/services/configuration/common/configurationModels.ts',
-			'src/vs/workbench/services/configuration/common/jsonEditingService.ts',
-			'src/vs/workbench/services/configuration/test/common/testServices.ts',
-			'src/vs/workbench/services/configurationResolver/browser/baseConfigurationResolverService.ts',
 			'src/vs/workbench/services/configurationResolver/common/configurationResolver.ts',
 			'src/vs/workbench/services/configurationResolver/common/configurationResolverExpression.ts',
-			'src/vs/workbench/services/configurationResolver/common/variableResolver.ts',
-			'src/vs/workbench/services/extensionManagement/browser/builtinExtensionsScannerService.ts',
-			'src/vs/workbench/services/extensionManagement/browser/webExtensionsScannerService.ts',
-			'src/vs/workbench/services/extensions/common/extHostCustomers.ts',
 			'src/vs/workbench/services/extensions/common/extensionHostManager.ts',
-			'src/vs/workbench/services/extensions/common/extensionHostProtocol.ts',
-			'src/vs/workbench/services/extensions/common/extensionHostProxy.ts',
 			'src/vs/workbench/services/extensions/common/extensionsRegistry.ts',
 			'src/vs/workbench/services/extensions/common/lazyPromise.ts',
 			'src/vs/workbench/services/extensions/common/polyfillNestedWorker.protocol.ts',
-			'src/vs/workbench/services/extensions/common/proxyIdentifier.ts',
 			'src/vs/workbench/services/extensions/common/rpcProtocol.ts',
-			'src/vs/workbench/services/extensions/electron-browser/cachedExtensionScanner.ts',
-			'src/vs/workbench/services/extensions/electron-browser/localProcessExtensionHost.ts',
 			'src/vs/workbench/services/extensions/worker/polyfillNestedWorker.ts',
 			'src/vs/workbench/services/keybinding/browser/keybindingService.ts',
 			'src/vs/workbench/services/keybinding/browser/keyboardLayoutService.ts',
 			'src/vs/workbench/services/keybinding/common/keybindingEditing.ts',
-			'src/vs/workbench/services/keybinding/common/keybindingIO.ts',
 			'src/vs/workbench/services/keybinding/common/keymapInfo.ts',
 			'src/vs/workbench/services/language/common/languageService.ts',
-			'src/vs/workbench/services/languageDetection/browser/languageDetectionWorker.protocol.ts',
 			'src/vs/workbench/services/outline/browser/outline.ts',
 			'src/vs/workbench/services/outline/browser/outlineService.ts',
 			'src/vs/workbench/services/preferences/common/preferences.ts',
 			'src/vs/workbench/services/preferences/common/preferencesModels.ts',
 			'src/vs/workbench/services/preferences/common/preferencesValidation.ts',
 			'src/vs/workbench/services/remote/common/tunnelModel.ts',
-			'src/vs/workbench/services/search/common/localFileSearchWorkerTypes.ts',
 			'src/vs/workbench/services/search/common/replace.ts',
 			'src/vs/workbench/services/search/common/search.ts',
 			'src/vs/workbench/services/search/common/searchExtConversionTypes.ts',
 			'src/vs/workbench/services/search/common/searchExtTypes.ts',
-			'src/vs/workbench/services/search/common/searchService.ts',
 			'src/vs/workbench/services/search/node/fileSearch.ts',
 			'src/vs/workbench/services/search/node/rawSearchService.ts',
 			'src/vs/workbench/services/search/node/ripgrepTextSearchEngine.ts',
-			'src/vs/workbench/services/search/worker/localFileSearch.ts',
-			'src/vs/workbench/services/terminal/common/embedderTerminalService.ts',
-			'src/vs/workbench/services/textMate/browser/backgroundTokenization/worker/textMateTokenizationWorker.worker.ts',
-			'src/vs/workbench/services/textMate/browser/backgroundTokenization/worker/textMateWorkerHost.ts',
-			'src/vs/workbench/services/textMate/browser/textMateTokenizationFeatureImpl.ts',
 			'src/vs/workbench/services/textMate/common/TMGrammarFactory.ts',
 			'src/vs/workbench/services/themes/browser/fileIconThemeData.ts',
 			'src/vs/workbench/services/themes/browser/productIconThemeData.ts',
-			'src/vs/workbench/services/themes/browser/workbenchThemeService.ts',
 			'src/vs/workbench/services/themes/common/colorThemeData.ts',
 			'src/vs/workbench/services/themes/common/plistParser.ts',
 			'src/vs/workbench/services/themes/common/themeExtensionPoints.ts',
 			'src/vs/workbench/services/themes/common/workbenchThemeService.ts',
-			'src/vs/workbench/services/userActivity/common/userActivityRegistry.ts',
-			'src/vs/workbench/services/userData/browser/userDataInit.ts',
-			'src/vs/workbench/services/userDataProfile/browser/userDataProfileInit.ts',
-			'src/vs/workbench/services/userDataSync/browser/userDataSyncInit.ts',
-			'src/vs/workbench/services/userDataSync/browser/userDataSyncWorkbenchService.ts',
-			'src/vs/workbench/services/userDataSync/common/userDataSync.ts',
 			'src/vs/workbench/test/browser/workbenchTestServices.ts',
 			'src/vs/workbench/test/common/workbenchTestServices.ts',
 			'src/vs/workbench/test/electron-browser/workbenchTestServices.ts',
@@ -1038,7 +787,6 @@ export default tseslint.config(
 			'src/vs/server/node/remoteAgentEnvironmentImpl.ts',
 			'src/vs/server/node/remoteExtensionHostAgentServer.ts',
 			'src/vs/server/node/remoteExtensionsScanner.ts',
-			'src/vs/server/node/remoteTerminalChannel.ts',
 			// Tests
 			'**/*.test.ts',
 			'**/*.integrationTest.ts'
@@ -1053,7 +801,7 @@ export default tseslint.config(
 			'@typescript-eslint/no-explicit-any': [
 				'warn',
 				{
-					'fixToUnknown': true
+					'fixToUnknown': false
 				}
 			]
 		}
@@ -1735,7 +1483,7 @@ export default tseslint.config(
 						'readline',
 						'stream',
 						'string_decoder',
-						'tas-client-umd',
+						'tas-client',
 						'tls',
 						'undici',
 						'undici-types',
@@ -1824,7 +1572,7 @@ export default tseslint.config(
 						'vs/base/~',
 						'vs/base/parts/*/~',
 						'vs/platform/*/~',
-						'tas-client-umd', // node module allowed even in /common/
+						'tas-client', // node module allowed even in /common/
 						'@microsoft/1ds-core-js', // node module allowed even in /common/
 						'@microsoft/1ds-post-js', // node module allowed even in /common/
 						'@xterm/headless' // node module allowed even in /common/
@@ -1942,7 +1690,7 @@ export default tseslint.config(
 							'when': 'test',
 							'pattern': 'vs/workbench/contrib/*/~'
 						}, // TODO@layers
-						'tas-client-umd', // node module allowed even in /common/
+						'tas-client', // node module allowed even in /common/
 						'vscode-textmate', // node module allowed even in /common/
 						'@vscode/vscode-languagedetection', // node module allowed even in /common/
 						'@vscode/tree-sitter-wasm', // type import
