@@ -88,6 +88,28 @@ export async function resolveElectronConfiguration(options: LaunchOptions): Prom
 	};
 }
 
+function findFilePath(root: string, path: string): string {
+	// First check if the path exists directly in the root
+	const directPath = join(root, path);
+	if (fs.existsSync(directPath)) {
+		return directPath;
+	}
+
+	// If not found directly, search through subdirectories
+	const entries = fs.readdirSync(root, { withFileTypes: true });
+
+	for (const entry of entries) {
+		if (entry.isDirectory()) {
+			const found = join(root, entry.name, path);
+			if (fs.existsSync(found)) {
+				return found;
+			}
+		}
+	}
+
+	throw new Error(`Could not find ${path} in any subdirectory`);
+}
+
 export function getDevElectronPath(): string {
 	const buildPath = join(root, '.build');
 	const product = require(join(root, 'product.json'));
@@ -113,7 +135,8 @@ export function getBuildElectronPath(root: string): string {
 			return join(root, product.applicationName);
 		}
 		case 'win32': {
-			const product = require(join(root, 'resources', 'app', 'product.json'));
+			const productPath = findFilePath(root, join('resources', 'app', 'product.json'));
+			const product = require(productPath);
 			return join(root, `${product.nameShort}.exe`);
 		}
 		default:
@@ -125,6 +148,10 @@ export function getBuildVersion(root: string): string {
 	switch (process.platform) {
 		case 'darwin':
 			return require(join(root, 'Contents', 'Resources', 'app', 'package.json')).version;
+		case 'win32': {
+			const packagePath = findFilePath(root, join('resources', 'app', 'package.json'));
+			return require(packagePath).version;
+		}
 		default:
 			return require(join(root, 'resources', 'app', 'package.json')).version;
 	}
