@@ -15,7 +15,7 @@ export function applyCoreTools(server: McpServer, appService: ApplicationService
 	// Playwright keeps using this as a start... maybe it needs some massaging
 	// server.tool(
 	// 	'vscode_automation_restart',
-	// 	'Restart Notepad# with optional workspace or folder and extra arguments',
+	// 	'Restart VS Code with optional workspace or folder and extra arguments',
 	// 	{
 	// 		workspaceOrFolder: z.string().optional().describe('Optional path to workspace or folder to open'),
 	// 		extraArgs: z.array(z.string()).optional().describe('Optional extra command line arguments')
@@ -34,7 +34,7 @@ export function applyCoreTools(server: McpServer, appService: ApplicationService
 
 	tools.push(server.tool(
 		'vscode_automation_stop',
-		'Stop the Notepad# application',
+		'Stop the VS Code application',
 		async () => {
 			const app = await appService.getOrCreateApplication();
 			await app.stopTracing(undefined, true);
@@ -51,7 +51,7 @@ export function applyCoreTools(server: McpServer, appService: ApplicationService
 	// This doesn't seem particularly useful
 	// server.tool(
 	// 	'vscode_automation_get_quality',
-	// 	'Get the quality/build type of Notepad# (Dev, Insiders, Stable, etc.)',
+	// 	'Get the quality/build type of VS Code (Dev, Insiders, Stable, etc.)',
 	// 	async () => {
 	// 		const info = {
 	// 			quality: app.quality,
@@ -115,7 +115,7 @@ export function applyCoreTools(server: McpServer, appService: ApplicationService
 	// Defer to Playwright's tool
 	// server.tool(
 	// 	'vscode_automation_send_keybinding',
-	// 	'Send a keybinding to Notepad# (e.g., ctrl+shift+p, cmd+s)',
+	// 	'Send a keybinding to VS Code (e.g., ctrl+shift+p, cmd+s)',
 	// 	{
 	// 		keybinding: z.string().describe('The keybinding to send (e.g., ctrl+shift+p, cmd+s, escape)'),
 	// 		waitSelector: z.string().optional().describe('Optional CSS selector to wait for after sending the keybinding')
