@@ -15,7 +15,7 @@ export function applyEditorTools(server: McpServer, appService: ApplicationServi
 	// Playwright can probably figure this one out
 	// server.tool(
 	// 	'vscode_automation_editor_open_file',
-	// 	'Open a file in the Notepad# editor through quick open',
+	// 	'Open a file in the VS Code editor through quick open',
 	// 	{
 	// 		fileName: z.string().describe('Name of the file to open (partial names work)')
 	// 	},
