@@ -2,7 +2,7 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import * as minimist from 'minimist';
+import minimist from 'minimist';
 
 const [, , ...args] = process.argv;
 export const opts = minimist(args, {
