@@ -44,7 +44,7 @@ export function setup(logger: Logger) {
 			});
 		});
 
-		it.skip('inserts/edits code cell', async function () {
+		it('inserts/edits code cell', async function () {
 			const app = this.app as Application;
 			await app.workbench.notebook.openNotebook();
 			await app.workbench.notebook.focusNextCell();
@@ -64,18 +64,21 @@ export function setup(logger: Logger) {
 			await app.workbench.notebook.waitForMarkdownContents('', '');
 		});
 
-		it.skip('moves focus as it inserts/deletes a cell', async function () {
+		it('moves focus as it inserts/deletes a cell', async function () {
 			const app = this.app as Application;
 			await app.workbench.notebook.openNotebook();
+			await app.workbench.notebook.focusFirstCell();
 			await app.workbench.notebook.insertNotebookCell('code');
 			await app.workbench.notebook.waitForActiveCellEditorContents('');
-			await app.workbench.notebook.stopEditingCell();
+			await app.workbench.notebook.waitForTypeInEditor('# added cell');
+			await app.workbench.notebook.focusFirstCell();
+			await app.workbench.notebook.insertNotebookCell('code');
+			await app.workbench.notebook.waitForActiveCellEditorContents('');
 			await app.workbench.notebook.deleteActiveCell();
-			await app.workbench.notebook.editCell();
-			await app.workbench.notebook.waitForTypeInEditor('## hello2!');
+			await app.workbench.notebook.waitForActiveCellEditorContents('# added cell');
 		});
 
-		it.skip('moves focus in and out of output', async function () { // TODO@rebornix https://github.com/microsoft/vscode/issues/139270
+		it('moves focus in and out of output', async function () { // TODO@rebornix https://github.com/microsoft/vscode/issues/139270
 			const app = this.app as Application;
 			await app.workbench.notebook.openNotebook();
 			// first cell is a code cell that already has output
