@@ -205,7 +205,7 @@ async function loadTests(opts) {
 		'creates a snapshot', // self-testing
 		'validates a snapshot', // self-testing
 		'cleans up old snapshots', // self-testing
-		'issue #149412: Notepad# hangs when bad semantic token data is received', // https://github.com/microsoft/vscode/issues/192440
+		'issue #149412: VS Code hangs when bad semantic token data is received', // https://github.com/microsoft/vscode/issues/192440
 		'issue #134973: invalid semantic tokens should be handled better', // https://github.com/microsoft/vscode/issues/192440
 		'issue #148651: VSCode UI process can hang if a semantic token with negative values is returned by language service', // https://github.com/microsoft/vscode/issues/192440
 		'issue #149130: vscode freezes because of Bracket Pair Colorization', // https://github.com/microsoft/vscode/issues/192440
@@ -304,7 +304,7 @@ async function loadTests(opts) {
 			const msg = [];
 			for (const error of errors) {
 				console.error(`Error: Test run should not have unexpected errors:\n${error}`);
-				msg.push(String(error))
+				msg.push(String(error));
 			}
 			assert.ok(false, `Error: Test run should not have unexpected errors:\n${msg.join('\n')}`);
 		}
@@ -464,7 +464,7 @@ async function runTests(opts) {
 	await loadTests(opts);
 
 	const runner = mocha.run(async () => {
-		await createCoverageReport(opts)
+		await createCoverageReport(opts);
 		ipcRenderer.send('all done');
 	});
 
